+++ app/src/hx711_plato.c
@@ -47,7 +47,7 @@ int8_t HX711_plato_Read(int32_t* data) {
 }
 
 
-float HX711_plato_GetWeight(int32_t lectura) {
+int32_t HX711_plato_GetWeight(int32_t lectura) {
    return (float)(lectura - valor_base) / factor_de_conversion;
 }
